Rename junit suites test to avoid confusion with test suites feature
