+++ subprojects/testing-jvm/src/integTest/groovy/org/gradle/testing/junit/AbstractJUnitSuitesIntegrationTest.groovy
@@ -19,7 +19,7 @@ package org.gradle.testing.junit
 import org.gradle.integtests.fixtures.DefaultTestExecutionResult
 import org.gradle.testing.fixture.AbstractTestingMultiVersionIntegrationTest
 
-abstract class AbstractJUnitTestSuitesIntegrationTest extends AbstractTestingMultiVersionIntegrationTest {
+abstract class AbstractJUnitSuitesIntegrationTest extends AbstractTestingMultiVersionIntegrationTest {
     abstract String getTestFrameworkSuiteDependencies()
     abstract String getTestFrameworkSuiteImports()
     abstract String getTestFrameworkSuiteAnnotations(String classes)
