+++ subprojects/testing-jvm/src/integTest/groovy/org/gradle/testing/junit/jupiter/JUnitJupiterSuitesIntegrationTest.groovy
@@ -18,12 +18,12 @@ package org.gradle.testing.junit.jupiter
 
 import org.gradle.integtests.fixtures.TargetCoverage
 import org.gradle.testing.fixture.JUnitCoverage
-import org.gradle.testing.junit.AbstractJUnitTestSuitesIntegrationTest
+import org.gradle.testing.junit.AbstractJUnitSuitesIntegrationTest
 
 import static org.gradle.testing.fixture.JUnitCoverage.JUNIT_JUPITER
 
 @TargetCoverage({ JUNIT_JUPITER })
-class JUnitJupiterTestSuitesIntegrationTest extends AbstractJUnitTestSuitesIntegrationTest implements JUnitJupiterMultiVersionTest {
+class JUnitJupiterSuitesIntegrationTest extends AbstractJUnitSuitesIntegrationTest implements JUnitJupiterMultiVersionTest {
     @Override
     String getTestFrameworkSuiteImports() {
         return """
